# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MovieFlix API",
        "description": "Essa API é responsável por gerenciar uma tabela de filmes do projeto backend MovieFlix.",
        "version": "1.0.0"
    },
    "paths": {
        "/movies": {
            "get": {
                "tags": ["Movie"],
                "summary": "Obter todos os filmes.",
                "responses": {
                    "200": {
                        "description": "Retorna uma lista de filmes."
                    }
                }
            },
            "post": {
                "tags": ["Movie"],
                "summary": "Cadastrar um filme.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MovieInput"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "O campo \"title\" é obrigatório, não pode estar vazio nem ser apenas \"string\"."
                    },
                    "201": {
                        "description": "Filme cadastrado com sucesso!"
                    },
                    "409": {
                        "description": "Já existe um filme cadastrado com esse título."
                    },
                    "500": {
                        "description": "Falha ao cadastrar o filme."
                    }
                }
            }
        },
        "/movies/{id}": {
            "put": {
                "tags": ["Movie"],
                "summary": "Atualizar um filme.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do filme a ser atualizado:",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MovieUpdateInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Filme atualizado com sucesso!"
                    },
                    "400": {
                        "description": "ID inválido ou título vazio/inválido."
                    },
                    "404": {
                        "description": "Filme não encontrado."
                    },
                    "409": {
                        "description": "Já existe um filme cadastrado com esse título."
                    },
                    "500": {
                        "description": "Falha ao atualizar o registro do filme."
                    }
                }
            },
            "delete": {
                "tags": ["Movie"],
                "summary": "Remover um filme",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do filme a ser removido:",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "ID inválido."
                    },
                    "200": {
                        "description": "Filme deletado com sucesso!"
                    },
                    "404": {
                        "description": "O filme não foi encontrado."
                    },
                    "500": {
                        "description": "Não foi possível remover o filme."
                    }
                }
            }
        },
        "/movies/{genreName}": {
            "get": {
                "tags": ["Movie"],
                "summary": "Filtrar filmes por gênero.",
                "parameters": [
                    {
                        "name": "genreName",
                        "in": "path",
                        "required": true,
                        "description": "Nome do gênero:",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filtragem do gênero realizada!"
                    },
                    "500": {
                        "description": "Falha ao fazer a filtragem do gênero."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Movie": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "ID do filme."
                    },
                    "title": {
                        "type": "string",
                        "description": "Título do filme."
                    },
                    "director": {
                        "type": "string",
                        "description": "Diretor do filme."
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lançamento do filme."
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "Gênero do filme."
                    },
                    "genres": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "ID do gênero."
                            },
                            "name": {
                                "type": "string",
                                "description": "Nome do gênero."
                            }
                        }
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "Idioma do filme."
                    },
                    "languages": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "ID do idioma."
                            },
                            "name": {
                                "type": "string",
                                "description": "Nome do idioma."
                            }
                        }
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Quantidade de Oscars que o filme recebeu."
                    }
                }
            },
            "MovieInput": {
                "type": "object",
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Título do filme."               
                    },
                    "director": {
                        "type": "string",
                        "description": "Diretor do filme."
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lançamento do filme."
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "ID do gênero do filme."
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "ID do idioma do filme."
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Quantidade de Oscars que o filme recebeu."
                    }
                }
            },
            "MovieUpdateInput": {
                "type": "object",
                "description": "Campos opcionais para atualização parcial do filme.",
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Título do filme."
                    },
                    "director": {
                        "type": "string",
                        "description": "Diretor do filme."
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lançamento do filme."
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "ID do gênero do filme."
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "ID do idioma do filme."
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Quantidade de Oscars que o filme recebeu."
                    }
                }
            }
        }
    }
}